requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dateutils==0.6.6
future==0.17.1
idna==2.8
importlib-metadata==0.19
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.1
parse==1.12.1
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.2
PySocks==1.7.0
pytest==5.0.1
python-telegram-bot==12.0.0b1
pytest==5.0.1
python-dateutil==2.8.0
python-telegram-bot==12.0.0b1
python-twitter==3.5
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
schedule==0.6.0
six==1.12.0
tornado==6.0.3
tweepy==3.8.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
Mastodon.py==1.4.6
python-linkedin-v2